+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed content encoding errors when serving media files.
+- Fixed key ID to actor ID transformation for microblog.pub.
 
 ## [2.17.1] - 2024-04-29